requirements.txt
feedparser==6.0.10
requests==2.31.0
openai>=1.3.5
pyyaml==6.0.1
python-dotenv==1.0.0
schedule==1.2.0
click==8.1.7